fix model location
